Clean up pipeline module
